requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
matplotlib==3.9.0
numpy>=2.0.0
gspread==6.0.0
google-auth==2.27.0
playsound==1.2.2

